image-gen-chat-app/requirements.txt
diffusers
transformers
accelerate
torch
Pillow
customtkinter 
datetime